requirements.txt
Flask==2.0.2
GitPython==3.1.14
boto3==1.18.69